add contrib directory for Google Native Client
Native Client lets you port C/C++ programs to run in the Chrome browser. The
various other files included in this commit support uploading Load81 to Google
App Engine and publishing it in the Chrome Web Store.

The major feature still missing from this port is file IO. The initial program
is hardcoded.