Avoid repetition mechanism for SDLK_UNKNOWN characters
There are special characters that can only be introduced by the combination
of several keys pressed asynchronously.

For instance, in Spanish keyboard, the sequence for typing character '^' is:
Shift + '^' + (space, vowel or '^'). There could be a pause between the
first two keys and the actual introduced character (i.e: ^, â, ê, etc).

By the time the actual character is introduced in the editor, the
dependent keys have already been released. In other editors, these type of
characters cannot be introduced repeteadly. SDL reports SDLK_UNKNOWN as
symbol for these characters.